+++ packages/core/mashroom/type-definitions/api.ts
@@ -424,7 +424,7 @@ export type MashroomServerConfig = {
     readonly httpsPort: number | null | undefined;
     readonly tlsOptions: TlsOptions | null | undefined;
     readonly trustProxy: boolean | number | string;
-    readonly xPowerByHeader: string | null | undefined;
+    readonly xPoweredByHeader: string | null | undefined;
     readonly serverRootFolder: string;
     readonly tmpFolder: string;
     readonly externalPluginConfigFileNames: Array<string>;